pyproject.toml
[project]
name = "openai-realtime-voice-chat-on-streamlit"
version = "0.1.0"
description = "This application enables real-time voice chat using Streamlit and OpenAI's Realtime API, featuring streamlit-webrtc for audio streaming without relying on the server's microphone."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "streamlit-webrtc~=0.47",
    "streamlit~=1.38.0", # streamlit_webrtc (0.47.9) doesn't work with 1.39.0 for some reason
    "openai~=1.55",
    "websockets~=14.1",
    "numpy~=2.1",
    "av~=12.3",
    "markdown2",
    "pyyaml"
  
]
